+++ CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- New `WindowsCertificates.trustedCertificate` resource to install X.509 certificates to Windows certificate stores. This is a pure .NET implementation equivalent to PowerShell's `Import-Certificate` command, with parameterizable certificate store locations.
+
 ## [0.3.0] - 2025-12-31
 ### Changed
 - Update the dependencies.
@@ -42,4 +46,4 @@ The following resource types are supported:
 [0.1.0]: https://github.com/ForNeVeR/Fabricator/releases/tag/v0.1.0
 [0.2.0]: https://github.com/ForNeVeR/Fabricator/compare/v0.1.0...v0.2.0
 [0.3.0]: https://github.com/ForNeVeR/Fabricator/compare/v0.2.0...v0.3.0
-[Unreleased]: https://github.com/ForNeVeR/Fabricator/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/ForNeVeR/Fabricator/compare/v0.3.0...HEAD